DRY refactor for shared tests components
